pyproject.toml
[tool.poetry]
name = "motorq"
version = "0.1.0"
description = ""
authors = ["Yash Rathi <yash.rathi.2k5@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.112.0"}
databases = {extras = ["postgresql"], version = "^0.9.0"}
sqlalchemy = "^2.0.32"
alembic = "^1.13.2"
python-dotenv = "^1.0.1"
pymongo = "^4.8.0"
pydantic-settings = "^2.4.0"
beanie = "^1.26.0"
requests = "^2.32.3"
httpx = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
